backend/chat-server/requirements.txt
aiohttp>=3.9.0
python-socketio>=5.10.0
redis[hiredis]>=5.0.0
asyncpg>=0.29.0
PyJWT>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-aiohttp>=1.0.0
